AddAuthorizeToDishAndRestaurantAndRating

+++ Restaurants.Application/Ratings/Commands/DeleteRating/DeleteRatingCommandHandler.cs
@@ -1,13 +1,16 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Ratings.Commands.DeleteRating
 {
     public class DeleteRatingCommandHandler(ILogger<DeleteRatingCommandHandler> logger,
-    IRatingsRepository ratingsRepository) : IRequestHandler<DeleteRatingCommand>
+    IRatingsRepository ratingsRepository,
+    IRatingAuthorizationService ratingAuthorizationService) : IRequestHandler<DeleteRatingCommand>
     {
         public async Task Handle(DeleteRatingCommand request, CancellationToken cancellationToken)
         {
@@ -15,6 +18,9 @@ namespace Restaurants.Application.Ratings.Commands.DeleteRating
 
             var rating = await ratingsRepository.GetByIdAsync(request.Id)
                 ?? throw new NotFoundException(nameof(Rating), request.Id.ToString());
+
+            if (!ratingAuthorizationService.Authorize(rating, ResourceOperation.Delete))
+                throw new ForbidException();
 
             await ratingsRepository.DeleteAsync(rating);
         }
